// frontend/lib/api/generated/model/transferenciaAlmacenDetalleIn.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * SIGA-UNMSM API
 * API del Sistema Integral de Gestión de Almacén y Abastecimiento Alimentario — UNMSM. Módulos activos: Autenticación, Usuarios, Almacenes, Catálogo Nutricional.
 * OpenAPI spec version: 0.1.0
 */

export interface TransferenciaAlmacenDetalleIn {
  producto_id: number;
  /** @exclusiveMinimum 0 */
  cantidad_enviada: number;
}